backend/requirements.txt
altgraph==0.17
amqp==2.6.0
asgiref==3.2.3
billiard==3.6.3.0
captionbot==0.1.5
celery==4.4.6
certifi==2019.11.28
chardet==3.0.4
Click==7.0
defusedxml==0.6.0
Django==3.0
django-taggit==1.2.0
djangorestframework==3.11.0
Flask==1.1.1
flower==0.9.5
future==0.18.2
humanize==2.5.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.3
kombu==4.6.11
Markdown==3.1.1
MarkupSafe==1.1.1
numpy==1.18.0
oauthlib==3.1.0
pandas==0.25.3
pefile==2019.4.18
Pillow==7.0.0
prometheus-client==0.8.0
psycopg2==2.8.4
PyInstaller==3.6
PyJWT==1.7.1
PySimpleGUI==4.20.0
python-dateutil==2.8.1
python3-openid==3.1.0
pytz==2019.3
pywin32-ctypes==0.2.0
requests==2.22.0
requests-oauthlib==1.3.0
selenium==3.141.0
six==1.13.0
social-auth-app-django==2.1.0
social-auth-core==3.2.0
sorl-thumbnail==12.5.0
sqlparse==0.3.0
tornado==6.0.4
urllib3==1.25.7
vine==1.3.0
virtualenv==16.7.9
Werkzeug==0.16.0
